// packages/ui/styledComponents/dataGrid/models/colDef/gridColDef.ts
export type ValueOptions = string | number | { value: any; label: string };